+++ src/Building.cpp
@@ -52,8 +52,6 @@ Building::Building(const Polygon_with_attr& poly)
           m_ptsPtr(std::make_shared<Point_set_3>()), m_hasFailed(false), m_reconSettings(nullptr) {}
         // 'true' here to check for polygon simplicity
 
-Building::~Building() = default;
-
 void Building::insert_point(const Point_3& pt) {
     m_ptsPtr->insert(pt);
 }
@@ -284,17 +282,26 @@ PointSet3Ptr Building::get_points() const {
     return m_ptsPtr;
 }
 
-void Building::get_cityjson_info(nlohmann::json& b) const {
-    b["type"] = "Building";
+std::string Building::get_lod() const {
+    return m_reconSettings->lod;
+}
+
+void Building::get_cityjson_cityobj_info(nlohmann::json& f) const {
+    f["type"] = "Building";
 //  b["attributes"];
 //    get_cityjson_attributes(b, _attributes);
 //    float hbase = z_to_float(this->get_height_base());
 //    float h = z_to_float(this->get_height());
 //    b["attributes"]["TerrainHeight"] = m_baseElevations.back(); // temp - will calculate avg for every footprint
-    b["attributes"]["measuredHeight"] = m_elevation - geomutils::avg(m_groundElevations[0]);
+    f["attributes"]["measuredHeight"] = m_elevation - geomutils::avg(m_groundElevations[0]);
 }
 
-void Building::get_cityjson_semantics(nlohmann::json& g) const { // Temp for checking CGAL mesh properties
+void Building::get_cityjson_geomobj_info(nlohmann::json& g) const {
+    g["type"] = "MultiSurface";
+    g["lod"] = this->get_lod();
+}
+
+void Building::get_cityjson_semantics(nlohmann::json& g) const {
     Face_property semantics;
     auto semanticsMap = m_mesh.property_map<face_descriptor, std::string>("f:semantics");
     if (semanticsMap.has_value()) {
@@ -302,20 +309,16 @@ void Building::get_cityjson_semantics(nlohmann::json& g) const { // Temp for che
     } else throw city4cfd_error("Semantic property map not found!");
 
     std::unordered_map<std::string, int> surfaceId;
-    surfaceId["RoofSurface"]   = 0; g["semantics"]["surfaces"][0]["type"] = "RoofSurface";
-    surfaceId["GroundSurface"] = 1; g["semantics"]["surfaces"][1]["type"] = "GroundSurface";
-    surfaceId["WallSurface"]   = 2; g["semantics"]["surfaces"][2]["type"] = "WallSurface";
+    surfaceId["RoofSurface"]   = 0; g["surfaces"][0]["type"] = "RoofSurface";
+    surfaceId["GroundSurface"] = 1; g["surfaces"][1]["type"] = "GroundSurface";
+    surfaceId["WallSurface"]   = 2; g["surfaces"][2]["type"] = "WallSurface";
 
     for (auto faceIdx : m_mesh.faces()) {
         auto it = surfaceId.find(semantics[faceIdx]);
         if (it == surfaceId.end()) throw city4cfd_error("Could not find semantic attribute!");
 
-        g["semantics"]["values"][faceIdx.idx()] = it->second;
+        g["values"][faceIdx.idx()] = it->second;
     }
-}
-
-std::string Building::get_cityjson_primitive() const {
-    return "MultiSurface";
 }
 
 TopoClass Building::get_class() const {