+++ src/ImportedBuilding.cpp
@@ -39,7 +39,6 @@
 #include <CGAL/Polygon_mesh_processing/triangulate_faces.h>
 #include <CGAL/Polygon_mesh_processing/compute_normal.h>
 
-
 int ImportedBuilding::noBottom = 0;
 
 ImportedBuilding::ImportedBuilding(std::unique_ptr<nlohmann::json>& buildingJson, PointSet3Ptr& importedBuildingPts)
@@ -266,8 +265,6 @@ ImportedBuilding::ImportedBuilding(Mesh& mesh)
     (*m_buildingJson)["geometry"].front()["boundaries"].push_back(geom);
     mesh.clear();
 }
-
-ImportedBuilding::~ImportedBuilding() = default;
 
 /*
  * Calculate building elevation without reconstruction.