+++ src/SurfaceLayer.h
@@ -32,19 +32,15 @@
 
 class SurfaceLayer : public PolyFeature {
 public:
-    SurfaceLayer();
     SurfaceLayer(const int outputLayerID);
-    SurfaceLayer(const nlohmann::json& poly);
     SurfaceLayer(const nlohmann::json& poly, const int outputLayerID);
-    SurfaceLayer(const Polygon_with_attr& poly);
     SurfaceLayer(const Polygon_with_attr& poly, const int outputLayerID);
-    ~SurfaceLayer();
+    ~SurfaceLayer() = default;
 
     void check_feature_scope(const Polygon_2& bndPoly);
 
-    virtual void        get_cityjson_info(nlohmann::json& b) const override;
-    virtual void        get_cityjson_semantics(nlohmann::json& g) const override;
-    virtual std::string get_cityjson_primitive() const override;
+    virtual void        get_cityjson_cityobj_info(nlohmann::json& f) const override;
+    virtual void        get_cityjson_geomobj_info(nlohmann::json& g) const override;
     virtual TopoClass   get_class() const override;
     virtual std::string get_class_name() const override;