+++ src/TopoFeature.h
@@ -36,16 +36,17 @@ public:
     TopoFeature();
     TopoFeature(std::string pid);
     TopoFeature(int outputLayerID);
-    virtual ~TopoFeature();
+    virtual ~TopoFeature() = default;
 
     static void          add_recon_region_output_layers(const int numLayers);
     static int           get_num_output_layers();
 
-    virtual void         get_cityjson_info(nlohmann::json& b) const;
-    virtual void         get_cityjson_semantics(nlohmann::json& g) const;
-    virtual std::string  get_cityjson_primitive() const;
     virtual TopoClass    get_class() const = 0;
     virtual std::string  get_class_name() const = 0;
+
+    virtual void         get_cityjson_cityobj_info(nlohmann::json& f) const;
+    virtual void         get_cityjson_geomobj_info(nlohmann::json& g) const;
+    virtual void         get_cityjson_semantics(nlohmann::json& g) const;
 
     Mesh&       get_mesh();
     const Mesh& get_mesh() const;