+++ src/io.cpp
@@ -31,6 +31,7 @@
 #include "TopoFeature.h"
 #include "Boundary.h"
 #include "Building.h"
+#include "SurfaceLayer.h"
 
 #include <CGAL/Polygon_mesh_processing/compute_normal.h>
 #include <CGAL/Polygon_mesh_processing/connected_components.h>
@@ -421,44 +422,75 @@ void IO::output_cityjson(const OutputFeaturesPtr& allFeatures) {
     nlohmann::json j;
 
     j["type"] = "CityJSON";
-    j["version"] = "1.0";
+    j["version"] = "2.0";
     j["metadata"] = {};
-    std::vector<double> bbox = Boundary::get_domain_bbox();
-    j["metadata"]["geographicalExtent"] = Boundary::get_domain_bbox();
-    j["metadata"]["referenceSystem"] = "urn:ogc:def:crs:EPSG::7415";
+    j["CityObjects"] = {};
+    j["vertices"] = {};
+
+    // CityJSON transform and scale are hardcoded to 0 and 0.001
+    j["transform"] = {};
+    j["transform"]["scale"] = {0.001, 0.001, 0.001};
+    j["transform"]["translate"] = {0, 0, 0};
+
+//    j["metadata"]["referenceSystem"] = "urn:ogc:def:crs:EPSG::7415"; // CRS is not defined due to transform by point of interest
     std::unordered_map<std::string, int> dPts;
+    std::vector<double> minMaxZ = {global::largnum, -global::largnum};
     for (auto& f : allFeatures) {
-        // Only Buildings and Terrain for now
-        if (f->get_class() != BUILDING && f->get_class() != TERRAIN) continue;
+        if (f->get_class() != BUILDING &&
+            f->get_class() != TERRAIN &&
+            f->get_class() != SURFACELAYER) continue;
         //-- Get feature info
         nlohmann::json b;
-        f->get_cityjson_info(b);
+        f->get_cityjson_cityobj_info(b);
 
         //-- Get feature geometry
         nlohmann::json g;
-        IO::get_cityjson_geom(f->get_mesh(), g, dPts, f->get_cityjson_primitive());
+        f->get_cityjson_geomobj_info(g);
+
+        IO::get_cityjson_geom(f->get_mesh(), g, dPts, minMaxZ);
 
         //-- Get feature semantics
-        f->get_cityjson_semantics(g);
+        nlohmann::json s;
+        f->get_cityjson_semantics(s);
+        if (!s.empty()) g["semantics"] = s;
 
         //-- Append to main json struct
         b["geometry"].push_back(g);
-        j["CityObjects"][f->get_id()] = b;
+
+        //-- Store terran and surface layers with the name, rest with the id
+        if (f->get_class() == TERRAIN ||
+            f->get_class() == SURFACELAYER) {
+            j["CityObjects"][Config::get().outputSurfaces[f->get_output_layer_id()]] = b;
+        } else {
+            std::string uniqueID = f->get_id();
+            int suffix = 0;
+            // check if the ID already exists and ensure unique ID
+            while (!j["CityObjects"][uniqueID].empty()) {
+                ++suffix;
+                uniqueID = f->get_id() + "_" + std::to_string(suffix);
+            }
+            j["CityObjects"][uniqueID] = b;
+            if (suffix > 0)
+                Config::write_to_log("Building ID: " + f->get_id() + " already exists. Stored as a new ID: " + uniqueID);
+        }
     }
+    auto bbox = Boundary::get_outer_bnd_bbox();
+    bbox[2] = minMaxZ[0];
+    bbox[5] = minMaxZ[1];
+    j["metadata"]["geographicalExtent"] = bbox;
 
     //-- Vertices - store them in a vector to quickly sort
-    std::vector<std::string> thepts;
-    thepts.resize(dPts.size());
+    std::vector<std::string> thepts(dPts.size());
     for (auto& p : dPts)
         thepts[p.second] = p.first;
     dPts.clear();
     for (auto& p : thepts) {
         std::vector<std::string> c;
         boost::split(c, p, boost::is_any_of(" "));
-        j["vertices"].push_back({std::stod(c[0], NULL), std::stod(c[1], NULL), std::stod(c[2], NULL) });
+        j["vertices"].push_back({std::stoi(c[0]), std::stoi(c[1]), std::stoi(c[2])});
     }
 
-    of.open(Config::get().outputFileName + ".json");
+    of.open(Config::get().outputFileName + ".city.json");
     of << j.dump() << std::endl;
 }
 
@@ -505,26 +537,34 @@ void IO::get_stl_pts(Mesh& mesh, std::string& fs) {
     }
 }
 
-void IO::get_cityjson_geom(const Mesh& mesh, nlohmann::json& g, std::unordered_map<std::string, int>& dPts,
-                           std::string primitive) {
-    g["type"] = primitive;
-    g["lod"] = "1.2"; //hardcoded for now
+void IO::get_cityjson_geom(const Mesh& mesh,
+                           nlohmann::json& g,
+                           std::unordered_map<std::string, int>& dPts,
+                           std::vector<double>& minMaxZ) {
+    // CityJSON transform and scale are hardcoded to 0 and 0.001
+    const double inverseScale = 1000.; // hardcoded scale for CityJSON is 0.001, so we are multiplying by 1000 to store it
+    const double inverseTranslate = 0.; // hardcoded translation
+
     g["boundaries"];
     for (auto face: mesh.faces()) {
         if (IO::is_degen(mesh, face)) continue;
         std::vector<int> tempPoly;
         tempPoly.reserve(3);
         for (auto index: CGAL::vertices_around_face(mesh.halfedge(face), mesh)) {
-            std::string pt = gen_key_bucket(mesh.point(index));
+            // store point as a string with CityJSON transformations
+            std::string pt = gen_key_bucket_int(mesh.point(index), inverseScale, inverseTranslate);
             auto it = dPts.find(pt);
             if (it == dPts.end()) {
-
                 tempPoly.push_back(dPts.size());
 
                 dPts[pt] = dPts.size();
             } else {
                 tempPoly.push_back(it->second);
             }
+
+            // get the z value for the bounding box
+            minMaxZ[0] = std::min(minMaxZ[0], mesh.point(index).z());
+            minMaxZ[1] = std::max(minMaxZ[1], mesh.point(index).z());
         }
         g["boundaries"].push_back({tempPoly});
     }
@@ -639,3 +679,16 @@ std::string IO::gen_key_bucket(const T& p) {
 //- Explicit template instantiation
 template std::string IO::gen_key_bucket<Point_3>(const Point_3& p);
 template std::string IO::gen_key_bucket<Vector_3>(const Vector_3& p);
+
+template<typename T>
+std::string IO::gen_key_bucket_int(const T& p, const double inverseScale, const double inverseTranslate) {
+    std::stringstream ss;
+    // round to integer before converting to string
+    ss << static_cast<int>(std::round((p.x() + inverseTranslate) * inverseScale)) << " "
+       << static_cast<int>(std::round((p.y() + inverseTranslate) * inverseScale)) << " "
+       << static_cast<int>(std::round((p.z() + inverseTranslate) * inverseScale));
+    return ss.str();
+}
+//- Explicit template instantiation
+template std::string IO::gen_key_bucket_int<Point_3>(const Point_3& p, const double inverseScale, const double inverseTransform);
+template std::string IO::gen_key_bucket_int<Vector_3>(const Vector_3& p, const double inverseScale, const double inverseTransform);